# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OPC UA Light Server - Control API",
    "description": "REST API for managing the OPC UA Light Server: address space, server lifecycle, security, S7 PLC connections, and certificate trust.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:3100", "description": "Local development" }
  ],
...
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "enum": ["VALIDATION_ERROR","DUPLICATE_ERROR","NOT_FOUND","UNAUTHORIZED","INTERNAL_ERROR","SERVICE_UNAVAILABLE","RUNTIME_ERROR","CERTIFICATE_NOT_FOUND","ACCESS_DENIED","CONFLICT"] },
              "message": { "type": "string" },
              "details": { "type": "array", "items": { "$ref": "#/components/schemas/ErrorDetail" } }
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": ["field", "message"],
        "properties": {
          "field": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "OpcUaDataType": {
        "type": "string",
        "enum": ["Boolean","Int16","Int32","Int64","UInt16","UInt32","UInt64","Float","Double","String","DateTime","ByteString"]
      },
      "OpcUaNode": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "namespaceId": { "type": "string", "format": "uuid" },
          "objectNodeId": { "type": "string", "format": "uuid", "nullable": true },
          "name": { "type": "string" },
          "dataType": { "$ref": "#/components/schemas/OpcUaDataType" },
          "initialValue": {},
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateNodeRequest": {
        "type": "object",
        "required": ["name", "namespaceId", "dataType"],
        "properties": {
          "name": { "type": "string" },
          "namespaceId": { "type": "string", "format": "uuid" },
          "objectNodeId": { "type": "string", "format": "uuid", "nullable": true },
          "dataType": { "$ref": "#/components/schemas/OpcUaDataType" },
          "initialValue": {},
          "description": { "type": "string" }
        }
      },
...
      "Namespace": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "uri": { "type": "string" },
          "nodeCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateNamespaceRequest": {
        "type": "object",
        "required": ["name", "uri"],
        "properties": {
          "name": { "type": "string" },
          "uri": { "type": "string" },
          "description": { "type": "string" }
        }
      },
...
      "CreateObjectNodeRequest": {
        "type": "object",
        "required": ["name", "namespaceId"],
        "properties": {
          "name": { "type": "string" },
          "namespaceId": { "type": "string", "format": "uuid" },
          "parentObjectNodeId": { "type": "string", "format": "uuid", "nullable": true }
        }
      },
...
      "LogEntry": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "level": { "type": "string", "enum": ["info", "warn", "error", "debug"] },
          "source": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "ApiKeyAuth": [] },
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/nodes": {
      "get": {
        "tags": ["Nodes"],
        "summary": "List all nodes",
        "security": [],
        "parameters": [
          { "name": "namespaceId", "in": "query", "schema": { "type": "string" }, "description": "Filter by namespace" }
        ],
        "responses": {
          "200": { "description": "Node list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/OpcUaNode" } } } } }
        }
      },
      "post": {
        "tags": ["Nodes"],
        "summary": "Create a node",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateNodeRequest" } } } },
        "responses": {
          "201": { "description": "Node created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OpcUaNode" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Duplicate name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/namespaces": {
      "get": {
        "tags": ["Namespaces"],
        "summary": "List all namespaces",
        "security": [],
        "responses": { "200": { "description": "Namespace list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Namespace" } } } } } }
      },
      "post": {
        "tags": ["Namespaces"],
        "summary": "Create a namespace",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateNamespaceRequest" } } } },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Namespace" } } } },
          "400": { "description": "Validation error" },
          "409": { "description": "Duplicate" }
        }
      }
    },
...
    "/api/object-nodes": {
      "post": {
        "tags": ["Object Nodes"],
        "summary": "Create an object node",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateObjectNodeRequest" } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation error" }, "409": { "description": "Duplicate" } }
      }
    },
...
    "/api/s7/logs": {
      "get": {
        "tags": ["S7"],
        "summary": "Get S7 connector logs",
        "security": [],
        "parameters": [{ "name": "since", "in": "query", "schema": { "type": "string" }, "description": "ISO timestamp filter" }],
        "responses": { "200": { "description": "Log entries" } }
      }
    },
...
    "/api/logs": {
      "get": {
        "tags": ["Logs"],
        "summary": "Get system log entries",
        "security": [],
        "parameters": [
          { "name": "since", "in": "query", "schema": { "type": "string" }, "description": "ISO timestamp filter" },
          { "name": "level", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated levels" },
          { "name": "source", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated sources" }
        ],
        "responses": { "200": { "description": "Log entries", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/LogEntry" } } } } } }
      }
    }
  }
}